requirements.txt
appdirs==1.4.4
asgiref==3.3.1
blis==0.2.4
certifi==2020.12.5
chardet==4.0.0
# ChatterBot==1.0.5 - local now
chatterbot-corpus==1.2.0
click==7.1.2
cymem==2.0.5
distlib==0.3.1
Django==3.1.6
#localtest only:
#django-sslserver==0.20
django-recaptcha==2.0.6
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
joblib==1.0.1
mathparse==0.1.2
murmurhash==1.0.5
mysqlclient==1.4.6
# nltk==3.6.2 - local now
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.2.0
Pint==0.17
plac==0.9.6
preshed==2.0.1
psycopg2-binary==2.8.6
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2021.1
PyYAML==3.13
regex==2021.4.4
requests==2.25.1
six==1.15.0
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.4.1
srsly==1.0.5
thinc==7.0.8
tqdm==4.60.0
urllib3==1.26.4
uWSGI==2.0.19.1
virtualenv==20.4.2
wasabi==0.8.2
# ./Chatterbot-1.0.5-local.tar.gz
# ./nltk-3.5.tar.gz
